Move createPncDisposalsFromResult folder into lib/results folder
