+++ packages/core/lib/results/createPncDisposalsFromResult/createPncDisposal.test.ts
