+++ packages/core/lib/results/createPncDisposalsFromResult/formatDateSpecifiedInResult.ts
