+++ packages/core/lib/results/createPncDisposalsFromResult/index.ts
