+++ packages/core/lib/results/createPncDisposalsFromResult/isAmountSpecifiedInResultValid.ts
