+++ packages/core/lib/results/createPncDisposalsFromResult/isDriverDisqualificationResult.ts
