+++ packages/core/phase3/lib/hearingDetails/createArrestHearingFromOffence.ts
@@ -2,8 +2,8 @@ import type { Offence } from "../../../types/AnnotatedHearingOutcome"
 import type { PncUpdateDataset } from "../../../types/PncUpdateDataset"
 import type { PncUpdateArrestHearing } from "../../types/HearingDetails"
 
-import formatDateSpecifiedInResult from "../../../lib/createPncDisposalsFromResult/formatDateSpecifiedInResult"
 import getOffenceCode from "../../../lib/offences/getOffenceCode"
+import formatDateSpecifiedInResult from "../../../lib/results/createPncDisposalsFromResult/formatDateSpecifiedInResult"
 import { PncUpdateType } from "../../types/HearingDetails"
 import getForceStationCode from "../getForceStationCode"
 import { preProcessOffenceReasonSequence } from "./preProcessOffenceReasonSequence"
