+++ packages/core/phase3/lib/pncUpdateRequestGenerators/penaltyHearingGenerator.ts
@@ -1,6 +1,6 @@
 import type PncUpdateRequestGenerator from "../../types/PncUpdateRequestGenerator"
 
-import formatDateSpecifiedInResult from "../../../lib/createPncDisposalsFromResult/formatDateSpecifiedInResult"
+import formatDateSpecifiedInResult from "../../../lib/results/createPncDisposalsFromResult/formatDateSpecifiedInResult"
 import { PncOperation } from "../../../types/PncOperation"
 import generateBasePncUpdateRequest from "../generateBasePncUpdateRequest"
 import getPncCourtCode from "../getPncCourtCode"
